+++ src/Tegnforklaring/MapLegend.js
